Category changes
Changed the category_list and the product_list tables, created a new table named product_category to include the fact that a book might have many genres. I inserted some datas in all these 3 tables and in users table. I updated all the sections of the website that included category in it. If in products page, we need to add a new book, I made a checkbox for the category section. also in maintanance i made the buttons (view details, edit,delete and add category) more presentable.